requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
astronomer-providers==1.17.3
apache-airflow-providers-amazon==8.6.0
Faker==19.3.1
astro-sdk-python==1.7.0
apache-airflow-providers-http
boto3
pandas
requests
apache-airflow-providers-common-sql
